extract datadog code to ExternalLoggerManager
